Revert Worker changes
