--- packages/yarnpkg-core/sources/worker-zip/convert.ts
@@ -1,25 +1,0 @@
-import {PortablePath, ZipFS}                    from '@yarnpkg/fslib';
-import {getLibzipPromise}                       from '@yarnpkg/libzip';
-
-import {extractArchiveTo, ExtractBufferOptions} from '../tgzUtils';
-
-export type ConvertToZipPayload = {
-  tmpFile: PortablePath;
-  tgz: Buffer;
-  opts: ExtractBufferOptions;
-};
-
-export async function convertToZipImpl(data: ConvertToZipPayload) {
-  const {opts, tgz, tmpFile} = data;
-  const {compressionLevel, ...bufferOpts} = opts;
-
-  const zipFs = new ZipFS(tmpFile, {create: true, libzip: await getLibzipPromise(), level: compressionLevel});
-
-  // Buffers sent through Node are turned into regular Uint8Arrays
-  const tgzBuffer = Buffer.from(tgz);
-  await extractArchiveTo(tgzBuffer, zipFs, bufferOpts);
-
-  zipFs.saveAndClose();
-
-  return data.tmpFile;
-}
